requirements.txt
streamlit==1.37.1
numpy==1.24.3
pandas==2.1.4
scipy==1.10.1
scikit-learn==1.5.1
imbalanced-learn==0.12.3
joblib==1.3.2
catboost==1.2.2
tensorflow-cpu==2.16.2
keras==3.11.3

seaborn==0.13.0
matplotlib==3.8.2
plotly==5.18.0
statsmodels==0.14.2

python-dateutil
pytz
requests
streamlit-js-eval
supabase